Move ttl into cache directory
